description_scene: Add logic to prevent repeats
Added a history which gets reset every number of people turns, I chose
to do this rather than a floating window because it avoids getting into
a preditable pattern. However, it does mean that the same person can
show up twice in a row. I decided it was a decent trade off.

Signed-off-by: Meghan Halton <meghalton@gmail.com>